requirements.txt
python>=3.8
torch>=1.9.0
numpy>=1.19.0
transformers>=4.0.0
Pillow>=8.0.0